Add HF trainer (#9)
* add training code

* add docs and train code is working

* get back to encoder_fn

* fixed the tie problem

* add peft_config to the Zip2ZipModel

* update train docs

* dtype fix in the transformer encoder

* nits

* fix dimension when applying the mask

* fix the mask computation

* inference works; remove redundant mask

* trainer is working; add a few scripts to check checkpoints

* add resume checkpoint logi, and improve checkpoint loading

* remove device, dtype attribute from CodebookManger; now infer directly from input to guarantee locality with input

* remove device, dtype attribute from Zip2ZipModel and infer from input tensor; style

---------

+++ src/zip2zip/nn/encoders/transformer.py
@@ -34,8 +34,8 @@ class Layer(nn.Module):
         self.post_mlp_layernorm = nn.LayerNorm(hidden_size)
 
     def __call__(self, x: torch.Tensor, mask: torch.Tensor) -> torch.Tensor:
-        x = self.post_attention_layernorm(x + self.attention(x, mask))
-        x = self.post_mlp_layernorm(x + self.mlp(x))
+        x = self.post_attention_layernorm(x + self.attention(x, mask)).to(x.dtype)
+        x = self.post_mlp_layernorm(x + self.mlp(x)).to(x.dtype)
         return x
